@@ -19,21 +20,26 @@ def keystoreProperties = new Properties()
 keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
 android {
-    compileSdkVersion compile_sdk_version as int
-    buildToolsVersion build_tools_version
+    compileSdkVersion 25
+    buildToolsVersion '27.0.3'
 
     defaultConfig {
-        minSdkVersion min_sdk_version as int
-        targetSdkVersion target_sdk_version as int
+        minSdkVersion 19
+        targetSdkVersion 25
         versionCode VERSION_CODE
         versionName VERSION_NAME
         applicationId PACKAGE_NAME
         manifestPlaceholders = [appName: "@string/app_name"]
+        vectorDrawables.useSupportLibrary = true
+    }
+
+    dataBinding {
+        enabled = true
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     signingConfigs {
