@@ -1,1 +1,1 @@
-package com.sakurafish.parrot.callconfirm.ui;import android.app.Dialog;import android.content.Context;import android.graphics.drawable.AnimationDrawable;import android.os.Bundle;import android.view.LayoutInflater;import android.view.View;import android.widget.ImageView;import com.sakurafish.parrot.callconfirm.R;public class IncoProgressDialog extends Dialog {    private Context mContext;    public IncoProgressDialog(final Context context) {        this(context, android.R.style.Theme_Translucent_NoTitleBar);        mContext = context;    }    public IncoProgressDialog(final Context context, final int theme) {        super(context, theme);    }    @Override    protected void onCreate(final Bundle savedInstanceState) {        super.onCreate(savedInstanceState);        final LayoutInflater inflater = LayoutInflater.from(mContext);        final View loadingV = inflater.inflate(R.layout.inco_progress_dialog, null);        setContentView(loadingV);        setCancelable(false);        final ImageView image = (ImageView) loadingV.findViewById(R.id.progress);        final AnimationDrawable anim = (AnimationDrawable) image.getBackground();        image.post(new Runnable() {            @Override            public void run() {                anim.start();            }        });    }}
+package com.sakurafish.parrot.callconfirm.ui;import android.app.Dialog;import android.content.Context;import android.graphics.drawable.AnimationDrawable;import android.os.Bundle;import android.view.LayoutInflater;import android.view.View;import android.view.ViewGroup;import android.widget.ImageView;import com.sakurafish.parrot.callconfirm.R;public class IncoProgressDialog extends Dialog {    private Context mContext;    public IncoProgressDialog(final Context context) {        this(context, android.R.style.Theme_Translucent_NoTitleBar);        mContext = context;    }    public IncoProgressDialog(final Context context, final int theme) {        super(context, theme);    }    @Override    protected void onCreate(final Bundle savedInstanceState) {        super.onCreate(savedInstanceState);        final LayoutInflater inflater = LayoutInflater.from(mContext);        final ViewGroup nullParent = null;        final View loadingV = inflater.inflate(R.layout.inco_progress_dialog, nullParent);        setContentView(loadingV);        setCancelable(false);        final ImageView image = (ImageView) loadingV.findViewById(R.id.progress);        final AnimationDrawable anim = (AnimationDrawable) image.getBackground();        image.post(() -> anim.start());    }}
