@@ -20,7 +20,7 @@ public class Pref {
         final SharedPreferences pref = getSharedPreferences(con);
         final SharedPreferences.Editor editor = pref.edit();
         editor.putString(key, value);
-        editor.commit();
+        editor.apply();
     }
 
     public static String getPrefString(final Context con, final String key) {
@@ -28,11 +28,16 @@ public class Pref {
         return pref.getString(key, null);
     }
 
+    public static String getPrefString(final Context con, final String key, final String defValue) {
+        final SharedPreferences pref = getSharedPreferences(con);
+        return pref.getString(key, defValue);
+    }
+
     public static void setPref(final Context con, final String key, boolean value) {
         final SharedPreferences pref = getSharedPreferences(con);
         final SharedPreferences.Editor editor = pref.edit();
         editor.putBoolean(key, value);
-        editor.commit();
+        editor.apply();
     }
 
     public static boolean getPrefBool(final Context con, final String key, final boolean defaultBool) {
@@ -44,7 +49,7 @@ public class Pref {
         final SharedPreferences pref = getSharedPreferences(con);
         final SharedPreferences.Editor editor = pref.edit();
         editor.putInt(key, value);
-        editor.commit();
+        editor.apply();
     }
 
     public static int getPrefInt(final Context con, final String key) {
@@ -61,7 +66,7 @@ public class Pref {
         final SharedPreferences pref = getSharedPreferences(con);
         final SharedPreferences.Editor editor = pref.edit();
         editor.putLong(key, value);
-        editor.commit();
+        editor.apply();
     }
 
     public static boolean isExistKey(final Context con, final String key) {
