@@ -3,6 +3,8 @@ package com.sakurafish.parrot.callconfirm.activity;
 import android.app.Fragment;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.ActivityInfo;
+import android.databinding.DataBindingUtil;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -10,14 +12,19 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 
+import com.google.android.gms.ads.AdRequest;
 import com.sakurafish.parrot.callconfirm.R;
-import com.sakurafish.parrot.callconfirm.web.WebFragment;
+import com.sakurafish.parrot.callconfirm.databinding.ActivityMainBinding;
+import com.sakurafish.parrot.callconfirm.fragment.WebFragment;
+import com.sakurafish.parrot.callconfirm.utils.AdsHelper;
 
-import static com.sakurafish.parrot.callconfirm.web.BaseWebFragment.EXTRA_TITLE;
-import static com.sakurafish.parrot.callconfirm.web.BaseWebFragment.EXTRA_URL;
+import static com.sakurafish.parrot.callconfirm.fragment.BaseWebFragment.EXTRA_TITLE;
+import static com.sakurafish.parrot.callconfirm.fragment.BaseWebFragment.EXTRA_URL;
 
 
 public class WebViewActivity extends AppCompatActivity {
+
+    private ActivityMainBinding binding;
 
     private Fragment mContent;
     private String mUrl;
@@ -39,7 +46,9 @@ public class WebViewActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+
+        setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_main);
 
         initialize(savedInstanceState);
     }
@@ -65,6 +74,10 @@ public class WebViewActivity extends AppCompatActivity {
                     .replace(R.id.content, mContent)
                     .addToBackStack("WebFragment").commit();
         }
+
+        // show AD banner
+        AdRequest adRequest = new AdsHelper(this).getAdRequest();
+        binding.adView.loadAd(adRequest);
     }
 
     @Override
