@@ -1,8 +1,10 @@
 
-package com.sakurafish.parrot.callconfirm.web;
+package com.sakurafish.parrot.callconfirm.config;
 
 public class WebConsts {
 
+    public static final String LOCAL_INFORMATION = "file:///android_asset/www/information.html";
+    public static final String LOCAL_HELP = "file:///android_asset/www/help.html";
     public static final String LOCAL_CREDIT = "file:///android_asset/www/credit.html";
     public static final String LOCAL_PRIVACY_POLICY = "file:///android_asset/www/privacy_policy.html";
 
