@@ -1,9 +1,10 @@
-package com.sakurafish.parrot.callconfirm.web;
+package com.sakurafish.parrot.callconfirm.fragment;
 
 import android.annotation.SuppressLint;
 import android.app.Fragment;
 import android.content.ActivityNotFoundException;
 import android.content.Intent;
+import android.databinding.DataBindingUtil;
 import android.graphics.Bitmap;
 import android.graphics.Picture;
 import android.net.Uri;
@@ -22,6 +23,7 @@ import android.webkit.WebView.PictureListener;
 import android.webkit.WebViewClient;
 
 import com.sakurafish.parrot.callconfirm.R;
+import com.sakurafish.parrot.callconfirm.databinding.FragmentWebBinding;
 import com.sakurafish.parrot.callconfirm.ui.IncoProgressDialog;
 import com.sakurafish.parrot.callconfirm.utils.Utils;
 
@@ -37,8 +39,9 @@ public class BaseWebFragment extends Fragment {
     public static final String EXTRA_URL = "url";
     public static final String EXTRA_TITLE = "title";
 
-    protected WebView mWebView;
     protected IncoProgressDialog mLoading;
+
+    protected FragmentWebBinding binding;
 
     final protected WebViewClient mWebViewClient = new WebViewClient() {
         @Override
@@ -71,10 +74,8 @@ public class BaseWebFragment extends Fragment {
                     final Intent intent = Intent.parseUri(url, 0);
                     startActivity(intent);
                     return true;
-                } catch (URISyntaxException e) {
+                } catch (URISyntaxException | ActivityNotFoundException e) {
                     // Intent schemeが不正.
-                } catch (ActivityNotFoundException e) {
-                    // 対象アプリがインストールされていない.
                 }
                 return true;
             }
@@ -83,10 +84,8 @@ public class BaseWebFragment extends Fragment {
                     final Intent intent = Intent.parseUri(url, Intent.URI_INTENT_SCHEME);
                     startActivity(intent);
                     return true;
-                } catch (URISyntaxException e) {
+                } catch (URISyntaxException | ActivityNotFoundException e) {
                     // Intent schemeが不正.
-                } catch (ActivityNotFoundException e) {
-                    // 対象アプリがインストールされていない.
                 }
                 return true;
             }
@@ -148,7 +147,9 @@ public class BaseWebFragment extends Fragment {
 
     @Override
     public View onCreateView(final LayoutInflater inflater, final ViewGroup container, final Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_web, null);
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_web, container, false);
+        return binding.getRoot();
+
     }
 
     @Override
@@ -162,47 +163,46 @@ public class BaseWebFragment extends Fragment {
         mUrl = getArguments().getString(EXTRA_URL);
 
         // Viewのセットアップ
-        mWebView = (WebView) getView().findViewById(R.id.webView);
-        final WebSettings settings = mWebView.getSettings();
+        final WebSettings settings = binding.webView.getSettings();
         settings.setJavaScriptEnabled(true);
         settings.setBuiltInZoomControls(true);
         settings.setSupportZoom(true);
         settings.setAppCacheEnabled(true);
         settings.setCacheMode(WebSettings.LOAD_CACHE_ELSE_NETWORK);
         settings.setGeolocationEnabled(true);
-        mWebView.setWebChromeClient(mWebChromeClient);
-        mWebView.setWebViewClient(mWebViewClient);
-        mWebView.setPictureListener(mPictureLisener);
-        mWebView.setVerticalScrollbarOverlay(true);
+        binding.webView.setWebChromeClient(mWebChromeClient);
+        binding.webView.setWebViewClient(mWebViewClient);
+        binding.webView.setPictureListener(mPictureLisener);
+        binding.webView.setVerticalScrollbarOverlay(true);
 
         Utils.logDebug("url:" + mUrl);
         if (mBundle != null) {
-            mWebView.restoreState(mBundle);
+            binding.webView.restoreState(mBundle);
         } else if (savedInstanceState != null) {
-            mWebView.restoreState(savedInstanceState);
+            binding.webView.restoreState(savedInstanceState);
         } else {
-            mWebView.loadUrl(mUrl);
+            binding.webView.loadUrl(mUrl);
         }
     }
 
     @Override
     public void onSaveInstanceState(final Bundle outState) {
         super.onSaveInstanceState(outState);
-        if (mWebView != null) {
-            mWebView.saveState(outState);
+        if (binding.webView != null) {
+            binding.webView.saveState(outState);
         }
     }
 
     public void saveWebView() {
         mBundle = new Bundle();
-        if (mWebView != null) {
-            mWebView.saveState(mBundle);
+        if (binding.webView != null) {
+            binding.webView.saveState(mBundle);
         }
     }
 
     public boolean onBackPressed() {
-        if (mWebView.canGoBack()) {
-            mWebView.goBack();
+        if (binding.webView.canGoBack()) {
+            binding.webView.goBack();
             return true;
         } else {
             return false;
@@ -214,7 +214,7 @@ public class BaseWebFragment extends Fragment {
     }
 
     private void goBrowser(final String url) {
-        mWebView.stopLoading();
+        binding.webView.stopLoading();
         try {
             final Uri uri = Uri.parse(url);
             final Intent intent = new Intent(Intent.ACTION_VIEW, uri);
@@ -233,9 +233,8 @@ public class BaseWebFragment extends Fragment {
 
     private void finalizeLayout() {
         dismissLoading();
-        if (mWebView != null) {
-            mWebView.stopLoading();
-            mWebView = null;
+        if (binding.webView != null) {
+            binding.webView.stopLoading();
         }
     }
 
